+++ odoo/local-src/enfinfidu_report/__openerp__.py
@@ -11,7 +11,6 @@
     'images': [],
     'depends': [
         'report',
-        'stock',
     ],
     'data': [
         # Reports